SignIn theme removed Closed #10
